// patentior-backend/types/patient_types.ts
import { Diagnose } from "./diagnoses_types";

export enum Gender {
    Male = 'male',
    Female = 'female',
    Other = 'other',
}

export interface BaseEntry {
    id: string;
    description: string;
    date: string;
    specialist: string;
    diagnosisCodes?: Array<Diagnose['code']>;
}

export enum HealthCheckRating {
    "Healthy" = 0,
    "LowRisk" = 1,
    "HighRisk" = 2,
    "CriticalRisk" = 3,
}

export enum Type {
    HealthCheck = 'HealthCheck',
    OccupationalHealthCare = 'OccupationalHealthcare',
    Hospital = 'Hospital',
}

interface HealthCheckEntry extends BaseEntry {
    type: Type.HealthCheck;
    healthCheckRating: HealthCheckRating;
}

interface OccupationalHealthCareEntry extends BaseEntry {
    type: Type.OccupationalHealthCare;
    employerName: string;
    sickLeave?: {
        startDate: string,
        endDate: string,
    }
}

interface HospitalEntry extends BaseEntry {
    type: Type.Hospital;
    discharge: {
        date: string;
        criteria: string;
    },
}

export type Entry = HospitalEntry | OccupationalHealthCareEntry | HealthCheckEntry;

export interface Patient {
    id: string;
    name: string;
    dateOfBirth: string;
    ssn: string;
    gender: Gender;
    occupation: string;
    entries: Entry[];
}

export type PublicPatient = Omit<Patient, 'ssn' | 'entries'>

export type NewPatient = Omit<Patient, 'id'>;